+++ packages/rum/src/rum.ts
@@ -3,6 +3,7 @@ import {
   Configuration,
   Context,
   ContextValue,
+  deepMerge,
   ErrorContext,
   ErrorMessage,
   getTimestamp,
@@ -18,7 +19,6 @@ import {
   ResourceKind,
   withSnakeCaseKeys,
 } from '@datadog/browser-core'
-import lodashMerge from 'lodash.merge'
 
 import { LifeCycle, LifeCycleEventType } from './lifeCycle'
 import { matchRequestTiming } from './matchRequestTiming'
@@ -153,17 +153,18 @@ export function startRum(
 ): Omit<RumGlobal, 'init'> {
   let globalContext: Context = {}
 
-  internalMonitoring.setExternalContextProvider(() =>
-    lodashMerge(
-      {
-        application_id: applicationId,
-        session_id: viewContext.sessionId,
-        view: {
-          id: viewContext.id,
+  internalMonitoring.setExternalContextProvider(
+    () =>
+      deepMerge(
+        {
+          application_id: applicationId,
+          session_id: viewContext.sessionId,
+          view: {
+            id: viewContext.id,
+          },
         },
-      },
-      globalContext
-    )
+        globalContext
+      ) as Context
   )
 
   const batch = startRumBatch(
@@ -233,7 +234,7 @@ function startRumBatch(
     configuration.batchBytesLimit,
     configuration.maxMessageSize,
     configuration.flushTimeout,
-    () => lodashMerge(withSnakeCaseKeys(rumContextProvider()), globalContextProvider()),
+    () => deepMerge(withSnakeCaseKeys(rumContextProvider()), globalContextProvider()) as Context,
     beforeUnloadCallback
   )
   return {
@@ -263,7 +264,7 @@ function trackView(lifeCycle: LifeCycle, upsertRumEvent: (event: RumViewEvent, k
           documentVersion: view.documentVersion,
         },
         view: {
-          loadingTime: view.loadingTime,
+          loadingTime: view.loadingTime ? msToNs(view.loadingTime) : undefined,
           loadingType: view.loadingType,
           measures: view.measures,
         },