+++ packages/rum/src/userActionCollection.ts
@@ -1,5 +1,5 @@
-import { Context, DOM_EVENT, generateUUID, noop } from '@datadog/browser-core'
-import { getElementContent } from './getElementContent'
+import { Context, DOM_EVENT, generateUUID } from '@datadog/browser-core'
+import { getActionNameFromElement } from './getActionNameFromElement'
 import { LifeCycle, LifeCycleEventType } from './lifeCycle'
 import { trackEventCounts } from './trackEventCounts'
 import { waitIdlePageActivity } from './trackPageActivities'
@@ -45,7 +45,11 @@ export function startUserActionCollection(lifeCycle: LifeCycle) {
     if (!(event.target instanceof Element)) {
       return
     }
-    newUserAction(lifeCycle, UserActionType.CLICK, getElementContent(event.target))
+    const name = getActionNameFromElement(event.target)
+    if (!name) {
+      return
+    }
+    newUserAction(lifeCycle, UserActionType.CLICK, name)
   }
 
   // New views trigger the cancellation of the current pending User Action